+++ socmed/src/app/home/components/all-posts/all-posts.component.html
@@ -5,12 +5,13 @@
         <ion-grid>
           <ion-row class="ion-align-items-center ion-justify-content-center">
             <ion-col size="auto" class="ion-padding">
-              <ion-avatar>
+              <ion-avatar style="cursor: pointer;" [routerLink]="['/','home', post.author.id]">
                 <ion-img [src]="$any(post).fullImagePath"></ion-img>
               </ion-avatar>
             </ion-col>
             <ion-col>
-              <div class="full-name"> {{ post.author.firstName }} {{ post.author.lastName }}</div>
+              <div style="cursor: pointer;" [routerLink]="['/','home', post.author.id]" class="full-name">
+                {{ post.author.firstName }} {{ post.author.lastName }}</div>
               <div class="role">Full Stack Developer</div>
               <div class="meta-data">
                 1d<ion-icon name="ellipse"></ion-icon> Edited
