+++ socmed/src/app/home/components/connection-profile/connection-profile.component.html
@@ -1,3 +1,4 @@
+<app-header></app-header>
 <p>
   connection-profile works!
 </p>
