+++ socmed/src/app/home/components/profile-summary/profile-summary.component.html
@@ -1,9 +1,10 @@
 <ion-card>
-  <div [style.background]="bannerColors.colorTwo" class="backdrop">
-    <div [style.background]="bannerColors.colorTwo" class="background-one"></div>
-    <div [style.background]="bannerColors.colorOne" class="half-ellipse"></div>
-    <div [style.background]="bannerColors.colorTwo" class="background-two"></div>
-    <div [style.background]="bannerColors.colorThree" class="background-three"></div>
+  <div [style.background]="this.bannerColorService.bannerColors.bannerColors.colorTwo" class="backdrop">
+    <div [style.background]="this.bannerColorService.bannerColors.bannerColors.colorTwo" class="background-one"></div>
+    <div [style.background]="this.bannerColorService.bannerColors.bannerColors.colorOne" class="half-ellipse"></div>
+    <div [style.background]="this.bannerColorService.bannerColors.bannerColors.colorTwo" class="background-two"></div>
+    <div [style.background]="this.bannerColorService.bannerColors.bannerColors.colorThree" class="background-three">
+    </div>
   </div>
   <ion-card-header class="ion-text-center">
     <input #fileUpload (change)="onFileSelect($event)" type="file" style="display: none">
