+++ socmed/src/app/home/services/banner-color.service.ts
@@ -1,9 +1,49 @@
 import { Injectable } from '@angular/core';
+import { Role } from 'src/app/auth/models/user.model';
+
+type BannerColors = {
+  colorOne: string;
+  colorTwo: string;
+  colorThree: string;
+};
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: 'root',
 })
 export class BannerColorService {
+  bannerColors: BannerColors = {
+    colorOne: '#a0b4b7',
+    colorTwo: '#dbe7e9',
+    colorThree: '#bfd3d6;',
+  };
 
-  constructor() { }
+  constructor() {}
+
+  getBannerColors(role: Role): BannerColors {
+    switch (role) {
+      case 'admin':
+        return {
+          colorOne: '#daa520',
+          colorTwo: '#f0e68c',
+          colorThree: '#fafad2',
+        };
+
+      case 'premium':
+        return {
+          colorOne: '#bc8f8f',
+          colorTwo: '#c09999',
+          colorThree: '#ddadaf',
+        };
+
+      case 'user':
+        return {
+          colorOne: '#a0b4b7',
+          colorTwo: '#dbe7e9',
+          colorThree: '#bfd3d6',
+        };
+
+      default:
+        return this.bannerColors;
+    }
+  }
 }
